requirements.txt
pymodbus
paho-mqtt
configparser

